# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VulnGuard AI - Smart Contract Vulnerability Scanner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ScanRequest {
  /** Solidity smart contract source code */
  code: string;
  /**
   * Optional contract name
   * @nullable
   */
  contractName?: string | null;
}

export type VulnerabilitySeverity =
  (typeof VulnerabilitySeverity)[keyof typeof VulnerabilitySeverity];

export const VulnerabilitySeverity = {
  CRITICAL: "CRITICAL",
  HIGH: "HIGH",
  MEDIUM: "MEDIUM",
  LOW: "LOW",
} as const;

export interface Vulnerability {
  id: number;
  type: string;
  severity: VulnerabilitySeverity;
  /** @nullable */
  line_number: number | null;
  title: string;
  description: string;
  technical_risk: string;
  /** @nullable */
  vulnerable_code: string | null;
  /** @nullable */
  fixed_code: string | null;
  recommendation: string;
}

export interface ScanResult {
  success: boolean;
  scanId: string;
  contract_name: string;
  total_vulnerabilities: number;
  risk_score: number;
  vulnerabilities: Vulnerability[];
  summary: string;
  analysis_time_ms: number;
  timestamp: string;
}

export interface GenerateFixRequest {
  vulnerability: Vulnerability;
  contractCode: string;
}

export interface GenerateFixResult {
  success: boolean;
  /** @nullable */
  original_code: string | null;
  fixed_code: string;
  explanation: string;
  resources: string[];
}

export interface ErrorResponse {
  error: string;
}