# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VulnGuard AI - Smart Contract Vulnerability Scanner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /report/{scanId}:
    get:
      operationId: getReport
      tags: [scan]
      summary: Get a scan report as JSON
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "404":
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    Vulnerability:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        line_number:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: string
        technical_risk:
          type: string
        vulnerable_code:
          type: ["string", "null"]
        fixed_code:
          type: ["string", "null"]
        recommendation:
          type: string
      required:
        - id
        - type
        - severity
        - line_number
        - title
        - description
        - technical_risk
        - vulnerable_code
        - fixed_code
        - recommendation
    ScanResult:
      type: object
      properties:
        success:
          type: boolean
        scanId:
          type: string
        contract_name:
          type: string
        total_vulnerabilities:
          type: integer
        risk_score:
          type: integer
        vulnerabilities:
          type: array
          items:
            $ref: "#/components/schemas/Vulnerability"
        summary:
          type: string
        analysis_time_ms:
          type: integer
        timestamp:
          type: string
      required:
        - success
        - scanId
        - contract_name
        - total_vulnerabilities
        - risk_score
        - vulnerabilities
        - summary
        - analysis_time_ms
        - timestamp
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error