# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VulnGuard AI - Smart Contract Vulnerability Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { Vulnerability } from "./vulnerability";

export interface ScanResult {
  success: boolean;
  scanId: string;
  contract_name: string;
  total_vulnerabilities: number;
  risk_score: number;
  vulnerabilities: Vulnerability[];
  summary: string;
  analysis_time_ms: number;
  timestamp: string;
}